# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CUP FICCT Backend API",
        "description": "API REST del sistema de admisión al CUP FICCT.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor del backend"
        }
    ],
    "paths": {
...
        "/api/applicant-admission/postulantes": {
            "get": {
                "tags": [
                    "Postulantes"
                ],
                "summary": "Listar postulantes",
                "operationId": "54f8a438dd36576f80e42a51e987c202",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de postulantes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Postulantes"
                ],
                "summary": "Registrar postulante",
                "operationId": "b658e8a5b32ea447eb1d98e1e9c7c13f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "documento",
                                    "nombres",
                                    "apellidos",
                                    "email",
                                    "fecha_nacimiento",
                                    "colegio"
                                ],
                                "properties": {
                                    "documento": {
                                        "type": "string",
                                        "example": "9876543"
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "example": "María José"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Quispe Vargas"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "mquispe@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "70000000"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2007-03-15"
                                    },
                                    "colegio": {
                                        "type": "string",
                                        "example": "Colegio Nacional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Postulante creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Postulante"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applicant-admission/postulantes/{postulante}": {
            "get": {
                "tags": [
                    "Postulantes"
                ],
                "summary": "Ver postulante",
                "operationId": "4b6fd998ff0780a4e70afa717076d57a",
                "parameters": [
                    {
                        "name": "postulante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Postulante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Postulante"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Postulantes"
                ],
                "summary": "Editar postulante",
                "operationId": "0142d8f6b208eab7e94c4fbf5310ce5d",
                "parameters": [
                    {
                        "name": "postulante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "documento": {
                                        "type": "string"
                                    },
                                    "nombres": {
                                        "type": "string"
                                    },
                                    "apellidos": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "colegio": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Postulante actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Postulante"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Postulantes"
                ],
                "summary": "Eliminar postulante",
                "operationId": "987fc6e99603209254748b8c19211dc2",
                "parameters": [
                    {
                        "name": "postulante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Postulante eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applicant-admission/postulantes/{postulante}/titulo": {
            "get": {
                "tags": [
                    "Postulantes"
                ],
                "summary": "Descargar título (redirige a URL firmada)",
                "operationId": "097ae079eacccbb3967721ef7d7478c3",
                "parameters": [
                    {
                        "name": "postulante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirige al archivo"
                    },
                    "404": {
                        "description": "Sin título"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Postulantes"
                ],
                "summary": "Subir título de bachiller (PDF/imagen)",
                "operationId": "2faaabab3478299d6247a811368e4f2a",
                "parameters": [
                    {
                        "name": "postulante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "titulo"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Título subido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Postulante"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Postulante": {
                "properties": {
                    "documento": {
                        "type": "string",
                        "example": "9876543"
                    },
                    "nombres": {
                        "type": "string",
                        "example": "María José"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Quispe Vargas"
                    },
                    "email": {
                        "type": "string",
                        "example": "mquispe@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "70000000"
                    },
                    "fecha_nacimiento": {
                        "type": "string",
                        "format": "date",
                        "example": "2007-03-15"
                    },
                    "colegio": {
                        "type": "string",
                        "example": "Colegio Nacional"
                    },
                    "ciudad": {
                        "type": "string",
                        "example": "Santa Cruz"
                    },
                    "titulo_bachiller_path": {
                        "type": "string",
                        "example": "titulos_bachiller/9876543.pdf",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}